+++ foundation/eclipselink.core.test/src/it/java/org/eclipse/persistence/testing/tests/queries/report/Scenario1_9b.java
@@ -55,7 +55,7 @@ protected void setup()  throws Exception
             throw new TestErrorException("ReportQuery test failed: The result size are different");
         }
         BigDecimal expected = (BigDecimal)((ReportQueryResult)expectedResults.firstElement()).getByIndex(0);
-        BigDecimal result = (BigDecimal)ConversionManager.getDefaultManager().convertObject(((ReportQueryResult)results.firstElement()).getByIndex(0), BigDecimal.class);
+        BigDecimal result = ConversionManager.getDefaultManager().convertObject(((ReportQueryResult)results.firstElement()).getByIndex(0), BigDecimal.class);
         if (!Helper.compareBigDecimals(expected, result)) {
             throw new TestErrorException("ReportQuery test failed: The results did not match (" + expected + ", " + result + ")");
         }
