+++ foundation/org.eclipse.persistence.oracle/src/main/java/org/eclipse/persistence/platform/database/oracle/OracleJDBC_10_1_0_2ProxyConnectionCustomizer.java
@@ -207,7 +207,7 @@ public class OracleJDBC_10_1_0_2ProxyConnectionCustomizer extends ConnectionCust
     protected void buildProxyProperties() {
         Object proxyTypeValue = session.getProperty(PersistenceUnitProperties.ORACLE_PROXY_TYPE);
         try {
-            proxyType = (Integer) session.getPlatform().getConversionManager().convertObject(proxyTypeValue, Integer.class);
+            proxyType = session.getPlatform().getConversionManager().convertObject(proxyTypeValue, Integer.class);
         } catch (ConversionException conversionException) {
             throw ValidationException.oracleJDBC10_1_0_2ProxyConnectorRequiresIntProxytype();
         }
