+++ moxy/org.eclipse.persistence.moxy/src/test/java/org/eclipse/persistence/testing/oxm/mappings/simpletypes/typetranslator/childelement/Phone.java
@@ -32,7 +32,7 @@ public class Phone  {
     }
 
     public String toString() {
-        return "Phone(number=" + (Integer) number + ")";
+        return "Phone(number=" + number + ")";
     }
 
     public boolean equals(Object object) {
